Refactor tournament admin routes
