+++ src/routes/mod.rs
@@ -1,3 +1,4 @@
 pub mod user_routes;
 pub mod general_routes;
 pub mod tournament_routes;
+pub mod tournament_admin_routes;
